# api-gateway/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TWZ Fire Extinguisher Management System API",
    "version": "1.0.0",
    "description": "Complete RESTful API for fire extinguisher management. Auth uses a 2-step flow: POST /api/auth/login (credentials) → receive userId → POST /api/auth/verify-otp (OTP from email) → receive JWT tokens.",
    "contact": { "name": "TWZ Ltd", "email": "support@twzltd.com" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin","inspector","user"] },
          "isActive": { "type": "boolean" },
          "mustChangePassword": { "type": "boolean" },
          "phone": { "type": "string" },
          "department": { "type": "string" },
          "lastLoginAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/api/users": {
      "post": {
        "tags": ["2. User Management"],
        "summary": "Create user (Admin only)",
        "description": "Admin creates a new user. A temporary password is auto-generated and emailed to the user. User must change it on first login.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["firstName","lastName","email","role"],
            "properties": {
              "firstName": { "type": "string", "minLength": 2 },
              "lastName": { "type": "string", "minLength": 2 },
              "email": { "type": "string", "format": "email" },
              "role": { "type": "string", "enum": ["admin","inspector","user"] },
              "phone": { "type": "string" },
              "department": { "type": "string" }
            }
          }}}
        },
        "responses": {
          "201": { "description": "User created. Temp password emailed.", "content": { "application/json": { "schema": {
            "type": "object",
            "properties": { "message": { "type": "string" }, "user": { "$ref": "#/components/schemas/User" } }
          }}}},
          "409": { "description": "Email already registered" }
        }
      },
      "get": {
        "tags": ["2. User Management"],
        "summary": "List all users (Admin only, paginated)",
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 20 } },
          { "in": "query", "name": "role", "schema": { "type": "string", "enum": ["admin","inspector","user"] } },
          { "in": "query", "name": "isActive", "schema": { "type": "boolean" } },
          { "in": "query", "name": "search", "schema": { "type": "string", "description": "Search name or email" } }
        ],
        "responses": {
          "200": { "description": "Paginated user list", "content": { "application/json": { "schema": {
            "type": "object",
            "properties": {
              "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
              "pagination": { "$ref": "#/components/schemas/Pagination" }
            }
          }}}}
        }
      }
    },
...
  }
}